requirements.txt
amqp==2.5.2
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
decorator==4.4.1
Django==2.2.10
django-celery-beat==1.6.0
django-cleanup==4.0.0
django-debug-toolbar==2.2
django-extensions==2.2.8
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-rest-swagger==2.1.2
django-timezone-field==3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
ecdsa==0.14.1
entrypoints==0.3
Faker==4.0.2
flake8==3.7.9
future==0.18.2
gnureadline==8.0.0
idna==2.8
importlib-metadata==1.5.0
ipython==7.12.0
ipython-genutils==0.2.0
ishell==0.1.7
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
kombu==4.6.7
MarkupSafe==1.1.1
mccabe==0.6.1
monobank-api==0.1.2
more-itertools==8.2.0
openapi-codec==1.3.2
packaging==20.3
parso==0.6.1
pexpect==4.8.0
phonenumbers==8.12.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
pytest-mock==3.0.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
style==1.1.0
text-unidecode==1.3
traitlets==4.3.3
update==0.0.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
zipp==2.2.0
